requirement.txt
pillow
torch
torchvision
pytorch-lightning
diffusers
torchmetrics
lpips
